Change runShellCommand so that it doesn't use the shell.
This fixes a potentially serious security hole in the handling of
the shell arguments, because the workspacePath and fileName were not
escaped and were just passed to the shell verbatim.

To avoid using the shell at all, we can use -quit on the find command
line, so there's no need for the pipe to head -n 1.