Search parent directories if the file is not found in the current workspace.
If the file that we're looking for is not found in the current workspace,
search the parent directory.  This will find the file if it is in a
separate repo nearby to the current project.  This is easier than
parsing the project's header and framework search paths, but hopefully
will find the file in most cases.

This recursion stops if we have searched two levels up, or if we've
reached the user's home directory.  This is to avoid performing
searches that are too costly.