Remove the maxRange mechanism in favor of checking editedMask instead.
This fixes a bug where we would fail to process the log if it was cleared
and was then shorter than the log that came before (because maxRange did
not get reset).

Checking editedMask means that we only respond to character changes, not
attribute changes, which is what maxRange was trying to do anyway.